+++ obhod-core/files/usr/lib/constants.sh
@@ -1,6 +1,6 @@
 # shellcheck disable=SC2034
 
-OBHOD_VERSION="0.3.1"
+OBHOD_VERSION="0.3.2"
 ## Common
 OBHOD_CONFIG="/etc/config/obhod"
 RESOLV_CONF="/etc/resolv.conf"